Split HDReflectionProbeEditor.Handles into separate files
